react-chart.jsの導入準備#74
