+++ app/controllers/statistics_controller.rb
@@ -1,7 +1,5 @@
 class StatisticsController < ApplicationController
-  before_action :set_items_includes_logs, only: %i[pie line]
-  # 将来的には、チャートだけの切り替えにすること。つまり JSでの実装
-  def pie; end
+  before_action :set_items_includes_logs, only: %i[index]
 
-  def line; end
+  def index; end
 end
