+++ app/javascript/components/charts/Pie.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+export default function Pie() {
+  return (
+    <div></div>
+  )
+}
