+++ app/javascript/packs/ChartPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import "../stylesheets/application.scss";
+import Stats from "../components/pages/Stats";
+
+const root = ReactDOM.createRoot(document.getElementById('app'));
+
+root.render(
+  <Stats />
+);
