+++ app/views/layouts/navbar_items/_logged_in.html.erb
@@ -1,7 +1,7 @@
 <%# ログインしているnavbar_items_partial %>
 
 <li class="nav-item">
-  <%= link_to "統計", stats_pie_path, class: "nav-link" %>
+  <%= link_to "統計", stats_path, class: "nav-link" %>
   </li>
 <li class="nav-item">
   <%# ソーシャル機能のあしがかり%>
