+++ app/views/statistics/index.html.erb
@@ -1,0 +1,4 @@
+<%= render 'stats_header' %>
+
+<div id="app"></div>
+<%= javascript_pack_tag "ChartPage" %>
