+++ config/routes.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  get '/stats/pie', to: 'statistics#pie'
-  get '/stats/line', to: 'statistics#line'
+  get '/stats', to: 'statistics#index'
 
   resources :items, only: %i[index show edit update]
   resources :logs, only: %i[create destroy]
@@ -23,12 +22,11 @@ Rails.application.routes.draw do
   root 'static_pages#home'
   post '/guest', to: 'users#guest'
 
-  namespace :charts do
-    get 'all_lines'
-    get 'all_pies'
-  end
-
   namespace :api do
     resources :items, only: %i[index update]
+    namespace :charts do
+      get 'all_lines'
+      get 'all_pies'
+    end
   end
 end
